requirements.txt
Flask
tensorflow
pandas
numpy
matplotlib
scikit-learn
Pillow
